Make bootstrapping reproducible & add unit test
Set random seed to make sampling of border units and subsequent
estimation of the spatial separation index (venables) reproducible.

Add unit test for calculating UCI with bootstrapping_border=True. The
goal of bootstrapping is to improve the estimate of the maximum spatial
separation index (venables) compared to the heuristic approach (of using
an equal distribution along the border of the area of interest). Test if
the bootstrapping approach succeeds in this regard.

Note that for small areas, e.g. the 3x3 polygon grid from the unit test,
this may not always be the case. For the given random seed it is though.